refactor: Separate cargo config from -release config
